toggleComplete function
The to do list is now fully functional. You can add to dos, check, uncheck and delete them.

I may leave this project here. Although is doesn't look very good, the purpose was to practice the functionality aspects which I have now done.

Maybe I'll come back later and add to this, but I'm thinking at the moment to move on to something else.